% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbaconfound-package.R
\docType{package}
\name{qbaconfound-package}
\alias{qbaconfound-package}
\title{qbaconfound: Monte Carlo Quantitative Bias Analysis for Unmeasured Confounding}
\description{
A flexible Monte Carlo quantitative bias analysis (QBA) for unmeasured confounding in observational studies, as described in Hughes et al. The substantive analysis may be a generalised linear model or a Cox proportional hazards model with a binary, continuous, or categorical exposure and measured confounders. The method allows for one or more binary or continuous unmeasured confounders that may be correlated with the measured confounders. Informative priors for a small number of bias parameters encode external information about the unmeasured confounders.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/remlapmot/qbaconfound}
  \item Report bugs at \url{https://github.com/remlapmot/qbaconfound/issues}
}

}
\author{
\strong{Maintainer}: Tom Palmer \email{remlapmot@hotmail.com}

Authors:
\itemize{
  \item Tom Palmer \email{remlapmot@hotmail.com}
  \item Rachael Hughes
  \item Emily Kawabata
}

}
\keyword{internal}
