% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders.R
\name{u_binary}
\alias{u_binary}
\title{Specify a binary unmeasured confounder}
\usage{
u_binary(coef_out, coef_exp, prevalence, prev_dist = c("uniform", "beta"))
}
\arguments{
\item{coef_out}{Length-2 numeric \code{c(mean, sd)} giving the normal prior for
the coefficient of the unmeasured confounder in the outcome model
(\eqn{\beta_U}). Set \code{sd = 0} for a point-mass (fixed value) prior.}

\item{coef_exp}{Normal prior for the coefficient(s) of the exposure in the
model for the unmeasured confounder (\eqn{\alpha_X}). For a single
exposure term, a length-2 numeric \code{c(mean, sd)}. For a categorical
exposure with several terms, a matrix with one row \code{c(mean, sd)} per
exposure term, or a list of such length-2 vectors.}

\item{prevalence}{Length-2 numeric giving the prior for the marginal
prevalence \eqn{\pi} of the binary unmeasured confounder. If
\code{prev_dist = "uniform"} (the default) this is \code{c(min, max)} of a uniform
prior; if \code{prev_dist = "beta"} this is \code{c(a, b)} of a beta prior.}

\item{prev_dist}{Prior family for the prevalence: \code{"uniform"} or \code{"beta"}.}
}
\value{
An object of class \code{qba_u} describing a binary unmeasured
confounder.
}
\description{
Describes the prior distributions of the bias parameters for a single binary
unmeasured confounder, for use in \code{\link[=qbaconfound]{qbaconfound()}}.
}
\details{
As for \code{\link[=u_continuous]{u_continuous()}} the coefficient of \code{U} in the outcome model
(\code{coef_out}) and the coefficient(s) of the exposure in the model for \code{U}
(\code{coef_exp}) are bias parameters. For a binary confounder the third bias
parameter is the marginal prevalence of \code{U} (\code{prevalence}, i.e.
\eqn{\pi}) rather than a residual standard deviation. A prevalence is
usually easier to elicit and more readily reported in the literature than
the intercept of a logistic model; the intercept needed to reproduce the
drawn prevalence is derived internally.
}
\examples{
u_binary(coef_out = c(0.7, 0.1), coef_exp = c(0.4, 0.05),
         prevalence = c(0.15, 0.25))
}
\seealso{
\code{\link[=u_continuous]{u_continuous()}}, \code{\link[=qbaconfound]{qbaconfound()}}
}
